EES-1319 Re-merge ChartDataSetsAddForm into ChartDataSetsConfiguration
